error

+++ routes/web.php
@@ -15,18 +15,46 @@ use App\Http\Controllers\CarController;
 |
 */
 
+//Route::middleware(['auth', 'verified', 'role:user'])->group(function () {
+//});
+//Route::middleware(['auth', 'verified', 'role:admin'])->group(function () {
+//});
+
+
+Route::middleware(['auth', 'verified', 'role:admin'])->group(function () {
+    // cars
+    Route::get('/admin/cars', [CarController::class, 'index'])->name('cars.index');
+    Route::get('/admin/cars/create', [CarController::class, 'create'])->name('cars.create');
+    Route::post('/admin/cars/create', [CarController::class, 'store'])->name('cars.store');
+    Route::get('/admin/cars/{car}', [CarController::class, 'show'])->name('cars.show');
+    Route::get('/admin/cars/{car}/edit', [CarController::class, 'edit'])->name('cars.edit');
+    Route::put('/admin/cars/{car}/edit', [CarController::class, 'update'])->name('cars.update');
+    Route::delete('/admin/cars/{car}', [CarController::class, 'destroy'])->name('cars.destroy');
+
+    // reservations
+    Route::get('/admin/reservations', [ReservationController::class, 'index'])->name('reservations.index');
+    Route::get('/admin/reservations/create', [ReservationController::class, 'create'])->name('reservations.create');
+    Route::post('/admin/reservations', [ReservationController::class, 'store'])->name('reservations.store');
+    Route::get('/admin/reservations/{reservation}', [ReservationController::class, 'show'])->name('reservations.show');
+    // Add more admin-specific routes here
+});
+
+
+
 Route::get('/', function () {
     return view('welcome');
 });
 
 // cars
-Route::get('/cars', [CarController::class, 'index'])->name('cars.index');
-Route::get('/cars/create', [CarController::class, 'create'])->name('cars.create');
-Route::post('/cars/create', [CarController::class, 'store'])->name('cars.store');
+//Route::get('/cars', [CarController::class, 'index'])->name('cars.index');
+//Route::get('/cars/create', [CarController::class, 'create'])->name('cars.create');
+//Route::post('/cars/create', [CarController::class, 'store'])->name('cars.store');
 Route::get('/cars/{car}', [CarController::class, 'show'])->name('cars.show');
-Route::get('/cars/{car}/edit', [CarController::class, 'edit'])->name('cars.edit');
-Route::put('/cars/{car}/edit', [CarController::class, 'update'])->name('cars.update');
-Route::delete('/cars/{car}', [CarController::class, 'destroy'])->name('cars.destroy');
+//Route::get('/cars/{car}/edit', [CarController::class, 'edit'])->name('cars.edit');
+//Route::put('/cars/{car}/edit', [CarController::class, 'update'])->name('cars.update');
+//Route::delete('/cars/{car}', [CarController::class, 'destroy'])->name('cars.destroy');
+
+
 //Route::resource('/cars', CarController::class);
 //Route::get('/cars/{car}', [CarController::class, 'create'])->name('cars.car');
 //Route::get('/cars/{car}/edit', [CarController::class, 'create'])->name('cars.edit');
@@ -37,10 +65,10 @@ Route::delete('/cars/{car}', [CarController::class, 'destroy'])->name('cars.dest
 // reservations
 //Route::resource('/reservations', ReservationController::class);
 // Listing Reservations (Index)
-Route::get('/reservations', [ReservationController::class, 'index'])->name('reservations.index');
-Route::get('/reservations/create', [ReservationController::class, 'create'])->name('reservations.create');
-Route::post('/reservations', [ReservationController::class, 'store'])->name('reservations.store');
-Route::get('/reservations/{reservation}', [ReservationController::class, 'show'])->name('reservations.show');
+//Route::get('/reservations', [ReservationController::class, 'index'])->name('reservations.index');
+//Route::get('/reservations/create', [ReservationController::class, 'create'])->name('reservations.create');
+//Route::post('/reservations', [ReservationController::class, 'store'])->name('reservations.store');
+//Route::get('/reservations/{reservation}', [ReservationController::class, 'show'])->name('reservations.show');
 
 
 // authentication
